data_cleaning_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-data_cleaning_env"
version = "0.1.0"
description = "Data Cleaning Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
openai = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.1.0",
]
plot = [
    "matplotlib>=3.7.0",
]
ui = [
    "gradio>=4.0.0",
    "matplotlib>=3.7.0",
]
report = [
    "python-docx>=1.1.0",
]

[project.scripts]
server = "data_cleaning_env.server.app:main"
baseline = "data_cleaning_env.baseline_inference:main"
render-plot = "data_cleaning_env.plot_artifacts:_entry"

[tool.setuptools]
include-package-data = true
packages = ["data_cleaning_env", "data_cleaning_env.server"]
package-dir = { "data_cleaning_env" = ".", "data_cleaning_env.server" = "server" }

[tool.setuptools.package-data]
data_cleaning_env = ["tasks/**/*.csv", "tasks/**/*.json"]